+++ src/main/java/roomescape/application/dto/request/ReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.application.dto;
+package roomescape.application.dto.request;
 
 import java.time.LocalTime;
 import roomescape.domain.ReservationTime;
