+++ src/main/java/roomescape/application/dto/response/ThemeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.application.dto;
+package roomescape.application.dto.response;
 
 import roomescape.domain.Theme;
